C casts replaced by C++ or qobject casts
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@880 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/src/qwt_scale_map.cpp
@@ -284,7 +284,7 @@ QRectF QwtScaleMap::invTransform( const QwtScaleMap &xMap,
 QDebug operator<<( QDebug debug, const QwtScaleMap &map )
 {
     debug.nospace() << "QwtScaleMap("
-        << map.transformation()->type()
+        << static_cast<int>( map.transformation()->type() )
         << ", s:" << map.s1() << "->" << map.s2()
         << ", p:" << map.p1() << "->" << map.p2()
         << ")";
